[backend] Move dimension-related entities to newaccess
