+++ backend/backend/src/main/java/de/aaaaaaah/velcom/backend/access/entities/Measurement.java
@@ -1,5 +1,6 @@
 package de.aaaaaaah.velcom.backend.access.entities;
 
+import de.aaaaaaah.velcom.backend.newaccess.dimensionaccess.entities.Dimension;
 import de.aaaaaaah.velcom.shared.util.Either;
 import java.util.Objects;
