+++ backend/backend/src/main/java/de/aaaaaaah/velcom/backend/data/runcomparison/RunComparator.java
@@ -1,9 +1,9 @@
 package de.aaaaaaah.velcom.backend.data.runcomparison;
 
-import de.aaaaaaah.velcom.backend.access.entities.Dimension;
 import de.aaaaaaah.velcom.backend.access.entities.Measurement;
 import de.aaaaaaah.velcom.backend.access.entities.MeasurementValues;
 import de.aaaaaaah.velcom.backend.access.entities.Run;
+import de.aaaaaaah.velcom.backend.newaccess.dimensionaccess.entities.Dimension;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.HashMap;
