+++ backend/backend/src/main/java/de/aaaaaaah/velcom/backend/newaccess/caches/AvailableDimensionsCache.java
@@ -2,8 +2,8 @@ package de.aaaaaaah.velcom.backend.newaccess.caches;
 
 import com.github.benmanes.caffeine.cache.Cache;
 import com.github.benmanes.caffeine.cache.Caffeine;
-import de.aaaaaaah.velcom.backend.access.entities.Dimension;
 import de.aaaaaaah.velcom.backend.newaccess.dimensionaccess.DimensionReadAccess;
+import de.aaaaaaah.velcom.backend.newaccess.dimensionaccess.entities.Dimension;
 import de.aaaaaaah.velcom.backend.newaccess.repoaccess.entities.RepoId;
 import java.util.ArrayList;
 import java.util.Collection;
