+++ backend/backend/src/main/java/de/aaaaaaah/velcom/backend/newaccess/dimensionaccess/DimensionReadAccess.java
@@ -3,7 +3,7 @@ package de.aaaaaaah.velcom.backend.newaccess.dimensionaccess;
 import static org.jooq.codegen.db.tables.Measurement.MEASUREMENT;
 import static org.jooq.codegen.db.tables.Run.RUN;
 
-import de.aaaaaaah.velcom.backend.access.entities.Dimension;
+import de.aaaaaaah.velcom.backend.newaccess.dimensionaccess.entities.Dimension;
 import de.aaaaaaah.velcom.backend.newaccess.repoaccess.entities.RepoId;
 import de.aaaaaaah.velcom.backend.storage.db.DBReadAccess;
 import de.aaaaaaah.velcom.backend.storage.db.DatabaseStorage;
