+++ backend/backend/src/main/java/de/aaaaaaah/velcom/backend/newaccess/dimensionaccess/entities/DimensionInfo.java
@@ -1,5 +1,6 @@
-package de.aaaaaaah.velcom.backend.access.entities;
+package de.aaaaaaah.velcom.backend.newaccess.dimensionaccess.entities;
 
+import de.aaaaaaah.velcom.backend.access.entities.Measurement;
 import java.util.Objects;
 
 public class DimensionInfo {
