+++ backend/backend/src/main/java/de/aaaaaaah/velcom/backend/restapi/exception/NoSuchDimensionException.java
@@ -1,6 +1,6 @@
 package de.aaaaaaah.velcom.backend.restapi.exception;
 
-import de.aaaaaaah.velcom.backend.access.entities.Dimension;
+import de.aaaaaaah.velcom.backend.newaccess.dimensionaccess.entities.Dimension;
 
 public class NoSuchDimensionException extends RuntimeException {
