+++ backend/backend/src/main/java/de/aaaaaaah/velcom/backend/restapi/jsonobjects/JsonMeasurement.java
@@ -1,9 +1,9 @@
 package de.aaaaaaah.velcom.backend.restapi.jsonobjects;
 
-import de.aaaaaaah.velcom.backend.access.entities.DimensionInfo;
 import de.aaaaaaah.velcom.backend.access.entities.Measurement;
 import de.aaaaaaah.velcom.backend.access.entities.MeasurementError;
 import de.aaaaaaah.velcom.backend.access.entities.MeasurementValues;
+import de.aaaaaaah.velcom.backend.newaccess.dimensionaccess.entities.DimensionInfo;
 import de.aaaaaaah.velcom.shared.util.Either;
 import java.util.List;
 import javax.annotation.Nullable;
